xchat: openssl fix
Primarily, a fix to shove xchat off of using the system OpenSSL.
Otherwise, general modernisation and passing the new strict audit.

I’ve also allowed the applications/pixmaps folders to be installed
again, since we allow plenty of that in the core, and they can actually
be rather useful for x11 applications to keep. I had a good scout
around the core for issues as to why they were removed, but there
didn’t seem to be any.

Closes #6.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>